TestMedicineService test case error fixed
