+++ src/test/java/in/nethaji/service/TestMedicineService.java
@@ -4,15 +4,17 @@ import static org.junit.Assert.*;
 
 import java.util.List;
 
+import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import in.nethaji.model.Medicine;
 
-public class TestDisplayMedicine {
-
+public class TestMedicineService {
+	
 	@BeforeClass
 	public static void setUp() throws Exception {
+		
 		List<Medicine> medicineList = MedicineService.getMedicineList();
 
 		Medicine medicine1 = new Medicine("Paracetamol", 10.0, "2021-10-10", 100);
@@ -23,6 +25,27 @@ public class TestDisplayMedicine {
 		medicineList.add(medicine2);
 	}
 
+	@Test
+	public void testAddMedicine1() {
+		try {
+			Medicine medicine = new Medicine("Paracetomal", 10, "2021-10-10", 20);
+			boolean isAdded = MedicineService.addMedicine(medicine);
+			assertTrue(isAdded);
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+	}
+
+	@Test
+	public void testAddMedicine2() {
+		try {
+			Medicine medicine = new Medicine(null, 20.0, "20-10-2022", 30);
+			MedicineService.addMedicine(medicine);
+		} catch (Exception e) {
+			assertEquals("Invalid Medicine Name", e.getMessage());
+		}
+	}
+	
 	/**
 	 * This method is used to test display patients
 	 */
@@ -30,7 +53,7 @@ public class TestDisplayMedicine {
 	@Test
 	public void testGetMedicine() {
 		List<Medicine> medicineList = MedicineService.getMedicineList();
-		assertEquals(2, medicineList.size());
+		assertEquals(3, medicineList.size());
 	}
 
 }
